pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.8",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
